--- a/dedent/dedent.go
+++ b/dedent/dedent.go
@@ -1,34 +1,73 @@
 package dedent
 
 import (
-	"strings"
+	"bytes"
 )
 
 // String automatically detects the maximum indentation shared by all lines and
 // trims them accordingly.
 func String(s string) string {
-	lines := strings.Split(s, "\n")
-	minIndent := -1
-
-	for _, l := range lines {
-		if len(l) == 0 {
-			continue
-		}
-
-		indent := len(l) - len(strings.TrimLeft(l, " "))
-		if minIndent == -1 || indent < minIndent {
-			minIndent = indent
-		}
-	}
-
-	if minIndent <= 0 {
+	indent := minIndent(s)
+	if indent == 0 {
 		return s
 	}
 
-	var buf strings.Builder
-	for _, l := range lines {
-		l = strings.TrimPrefix(l, strings.Repeat(" ", minIndent))
-		buf.WriteString(l + "\n")
+	return dedent(s, indent)
+}
+
+func minIndent(s string) int {
+	var (
+		curIndent    int
+		minIndent    int
+		shouldAppend = true
+	)
+	for i := 0; i < len(s); i++ {
+		switch s[i] {
+		case ' ':
+			if shouldAppend {
+				curIndent++
+			}
+		case '\n':
+			curIndent = 0
+			shouldAppend = true
+		default:
+			if curIndent > 0 && (minIndent == 0 || curIndent < minIndent) {
+				minIndent = curIndent
+				curIndent = 0
+			}
+			shouldAppend = false
+		}
 	}
-	return strings.TrimSuffix(buf.String(), "\n")
+
+	return minIndent
+}
+
+func dedent(s string, indent int) string {
+	var (
+		omitted    int
+		shouldOmit = true
+		buf        bytes.Buffer
+	)
+
+	for i := 0; i < len(s); i++ {
+		switch s[i] {
+		case ' ':
+			if shouldOmit {
+				if omitted < indent {
+					omitted++
+					continue
+				}
+				shouldOmit = false
+			}
+			buf.WriteByte(s[i])
+		case '\n':
+			omitted = 0
+			shouldOmit = true
+			buf.WriteByte(s[i])
+		default:
+			buf.WriteByte(s[i])
+		}
+	}
+
+	return buf.String()
 }