@kiyonlin 2020-10-20
Support '\t':
```hs
BenchmarkDedent-8       13678540                90.1 ns/op            95 B/op          1 allocs/op
BenchmarkDedent-8       11221939                93.7 ns/op            95 B/op          1 allocs/op
BenchmarkDedent-8       12120711                93.9 ns/op            95 B/op          1 allocs/op
BenchmarkDedent-8       12527490                95.8 ns/op            95 B/op          1 allocs/op
```